requirements.txt
argh==0.26.2
certifi==2019.11.28
cffi==1.14.0
cryptography==2.8
decorator==4.4.2
future==0.18.2
pathtools==0.1.2
pycparser==2.20
python-telegram-bot==12.4.2
PyYAML==5.3.1
random-user-agent==1.0.1
selenium==3.141.0
six==1.14.0
tornado==6.0.4
urllib3==1.25.8
watchdog==0.10.2
